refactor: global error code types => contants 로 파일 위치 변경
