+++ src/coupon/presentation/http/filters/coupon-exception.filter.ts
@@ -19,7 +19,7 @@ import {
   CouponNotFoundError,
   UserCouponNotFoundError,
 } from "@/coupon/application/coupon.application.exceptions";
-import { ErrorCode } from "@/common/types/error-codes.enum";
+import { ErrorCode } from "@/common/constants/error-codes.enum";
 import {
   BaseExceptionFilter,
   ErrorMapping,
